Only log the aggregation of the keypresses once per minute
This improves security a lot, as we'll never be able to know
what exactly we pressed, so when typing passwords etc they
will not be stored.

This also makes the log files smaller, as we dont log
every keypress, which we don't even care about, we only
care about the patterns, so writing the number of presses
per key every minute is good enough for us.